+++ MarsEdu/Server/mars-edu-web/src/main/java/com/mars/edu/web/model/orm/SchoolGradeEntity.java
@@ -15,11 +15,11 @@ public class SchoolGradeEntity {
     private int schoolId;
     private String code;
     private String name;
-    private byte level;
+    private int level;
     private Integer managerId;
     private Integer leaderId;
     private String description;
-    private byte status;
+    private int status;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -64,11 +64,11 @@ public class SchoolGradeEntity {
 
     @Basic
     @Column(name = "level", nullable = false)
-    public byte getLevel() {
+    public int getLevel() {
         return level;
     }
 
-    public void setLevel(byte level) {
+    public void setLevel(int level) {
         this.level = level;
     }
 
@@ -104,11 +104,11 @@ public class SchoolGradeEntity {
 
     @Basic
     @Column(name = "status", nullable = false)
-    public byte getStatus() {
+    public int getStatus() {
         return status;
     }
 
-    public void setStatus(byte status) {
+    public void setStatus(int status) {
         this.status = status;
     }
